refactor(market-data): Phase5a extract track-worker and geyser_sync modules
Extract md-track-worker loop, Geyser coalesced push, and NATS coalescers
from market_data.rs into src/market_data/track/{worker,geyser_sync,coalesce}.rs.

- TrackWorkerContext trait wires MarketDataContext from the bin
- No behavior change: 1 OS thread, bounded enqueue, 500ms coalesce (I-4b)
- phase5a_track_worker_in_dedicated_module test + BUGS_FIXES HYBRID-PHASE5a

LOC: market_data.rs 20319 -> 19903 (-416); new track modules +678
