+++ src/market_data/track/mod.rs
@@ -1,6 +1,28 @@
+pub mod coalesce;
 pub mod desired_set;
+pub mod geyser_sync;
+pub mod worker;
 
+pub use coalesce::{
+    arb_coalesce_try_send, merge_arb_track_requests_updates, merge_momentum_active_pools_updates,
+    momentum_coalesce_try_send, spawn_arb_tracking_coalescer, spawn_momentum_tracking_coalescer,
+    MARKET_DATA_ARB_COALESCE_CHANNEL_CAP, MARKET_DATA_MOMENTUM_COALESCE_CHANNEL_CAP,
+};
 pub use desired_set::{
     pin_priority_from_consumer, symmetric_diff, ConsumerId, DesiredExplicitSet, ExplicitEntry,
     PinPriority,
+};
+pub use geyser_sync::{
+    consumer_id_for_track_pin, explicit_subscription_has_new_keys,
+    rebuild_desired_explicit_set_from_ctx, track_worker_execute_coalesced_push,
+    MARKET_DATA_MD_STATE_FLUSH_BUDGET_MS,
+};
+pub use worker::{
+    apply_arb_track_requests_on_track_worker, apply_momentum_active_pools_on_track_worker,
+    spawn_inline_track_worker_sender, spawn_noop_track_worker_sender, spawn_track_worker,
+    track_worker_process_command, track_worker_try_enqueue, TrackPinReason, TrackWorkerCommand,
+    TrackWorkerContext, TrackWorkerSender, MARKET_DATA_ARB_APPLY_CHUNK_SIZE,
+    MARKET_DATA_ARB_APPLY_CHUNK_THRESHOLD, MARKET_DATA_MOMENTUM_APPLY_CHUNK_SIZE,
+    MARKET_DATA_MOMENTUM_APPLY_CHUNK_THRESHOLD, MARKET_DATA_TRACK_WORKER_COALESCE_MS,
+    MARKET_DATA_TRACK_WORKER_QUEUE_CAP,
 };
